Speedup, bug fix: no neighbors' lists del alloc! See details.
A loop in the birth_coord function intended to remove an allocated coordinate from empty coordinate lists of neighboring coordinates wasn't looking for and deleting the allocated coordinate (it was looking for something else which never would be in those lists)! Fixed (in now looks for and deletes the allocated coordinate). Methinks this bug has been here from the start :(

Fixing this bug itself provides a substantial speed improvement, as coords not intended to be reused were being attempted to be reused (although I think the attempts were correctly, silently failing)--now they can't even be attempted to be reused.

Previous test render average: 5.8121s. Post bug-fix: 4.2368s!